Add header probabilites and first attempt at classifying headers
Currently not robust enough as unseen words not in the word bag
will return most common label, (i.e. company or id it seems)
and as there are a lot not yet classified this is seriously
throwing off which header is closest. I don't think there's
anything seriously wrong with the implementation and I think
that the issue could be fixed with more labeled training data.